# workspace/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for Matriz de Riesgos LAFT
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /riesgos:
    get:
      operationId: getRiesgos
      tags: [riesgos]
      summary: List all risks
      parameters:
        - name: proceso
          in: query
          schema: { type: string }
        - name: perfil
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Riesgo"
    post:
      operationId: createRiesgo
      tags: [riesgos]
      summary: Create a new risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiesgoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Riesgo"
  /riesgos/{id}:
    get:
      operationId: getRiesgo
      tags: [riesgos]
      summary: Get a risk by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Risk detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiesgoDetalle"
        "404":
          description: Not found
    patch:
      operationId: updateRiesgo
      tags: [riesgos]
      summary: Update a risk
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiesgoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Riesgo"
    delete:
      operationId: deleteRiesgo
      tags: [riesgos]
      summary: Delete a risk
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── CONTROLES ─────────────────────────────────────────────────────────────────
...
  /riesgos/{riesgoId}/controles:
    get:
      operationId: getRiesgoControles
      tags: [controles]
      summary: Get controls for a specific risk
      parameters:
        - name: riesgoId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Controls for the risk
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiesgoControlDetalle"
    post:
      operationId: addControlToRiesgo
      tags: [controles]
      summary: Associate a control with a risk
      parameters:
        - name: riesgoId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiesgoControlInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiesgoControlDetalle"
  /riesgos/{riesgoId}/controles/{controlId}:
    delete:
      operationId: removeControlFromRiesgo
      tags: [controles]
      summary: Remove a control from a risk
      parameters:
        - name: riesgoId
          in: path
          required: true
          schema: { type: integer }
        - name: controlId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── MONITOREO ─────────────────────────────────────────────────────────────────
...
  /parametros:
    get:
      operationId: getParametros
      tags: [parametros]
      summary: Get all parameters (control weights)
      responses:
        "200":
          description: List of parameters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parametro"
    post:
      operationId: createParametro
      tags: [parametros]
      summary: Create a parameter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParametroInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parametro"
  /parametros/{id}:
    patch:
      operationId: updateParametro
      tags: [parametros]
      summary: Update a parameter
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParametroUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Parametro"

  # ── DASHBOARD / WOW ENDPOINTS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Riesgo:
      type: object
      required:
        - id
        - codigo
        - proceso
        - descripcion
        - perfilInherente
        - perfilResidual
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        codigo: { type: string }
        proceso: { type: string }
        subproceso: { type: ["string", "null"] }
        factorRiesgo: { type: ["string", "null"] }
        descripcion: { type: string }
        riesgoLaft: { type: boolean }
        riesgoOperativo: { type: boolean }
        riesgoLegal: { type: boolean }
        riesgoReputacional: { type: boolean }
        riesgoContagio: { type: boolean }
        quePuedeSuceder: { type: ["string", "null"] }
        tipologia: { type: ["string", "null"] }
        porQuePuedeSuceder: { type: ["string", "null"] }
        consecuencias: { type: ["string", "null"] }
        probabilidadInherente: { type: ["number", "null"] }
        impactoInherente: { type: ["number", "null"] }
        perfilInherente: { type: string }
        probabilidadResidual: { type: ["number", "null"] }
        impactoResidual: { type: ["number", "null"] }
        perfilResidual: { type: string }
        promedioEfectividad: { type: ["number", "null"] }
        tipoMonitoreo: { type: ["string", "null"] }
        responsableMonitoreo: { type: ["string", "null"] }
        periodicidadMonitoreo: { type: ["string", "null"] }
        prioridad: { type: ["string", "null"] }
        sugerencias: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    RiesgoInput:
      type: object
      required: [codigo, proceso, descripcion]
      properties:
        codigo: { type: string }
        proceso: { type: string }
        subproceso: { type: string }
        factorRiesgo: { type: string }
        descripcion: { type: string }
        riesgoLaft: { type: boolean }
        riesgoOperativo: { type: boolean }
        riesgoLegal: { type: boolean }
        riesgoReputacional: { type: boolean }
        riesgoContagio: { type: boolean }
        quePuedeSuceder: { type: string }
        tipologia: { type: string }
        porQuePuedeSuceder: { type: string }
        consecuencias: { type: string }
        probabilidadInherente: { type: number }
        impactoInherente: { type: number }
        probabilidadResidual: { type: number }
        impactoResidual: { type: number }
        tipoMonitoreo: { type: string }
        responsableMonitoreo: { type: string }
        periodicidadMonitoreo: { type: string }
        prioridad: { type: string }
        sugerencias: { type: string }
    RiesgoUpdate:
      type: object
      properties:
        codigo: { type: string }
        proceso: { type: string }
        subproceso: { type: string }
        factorRiesgo: { type: string }
        descripcion: { type: string }
        riesgoLaft: { type: boolean }
        riesgoOperativo: { type: boolean }
        riesgoLegal: { type: boolean }
        riesgoReputacional: { type: boolean }
        riesgoContagio: { type: boolean }
        quePuedeSuceder: { type: string }
        tipologia: { type: string }
        porQuePuedeSuceder: { type: string }
        consecuencias: { type: string }
        probabilidadInherente: { type: number }
        impactoInherente: { type: number }
        probabilidadResidual: { type: number }
        impactoResidual: { type: number }
        tipoMonitoreo: { type: string }
        responsableMonitoreo: { type: string }
        periodicidadMonitoreo: { type: string }
        prioridad: { type: string }
        sugerencias: { type: string }
    RiesgoDetalle:
      allOf:
        - $ref: "#/components/schemas/Riesgo"
        - type: object
          properties:
            controles:
              type: array
              items:
                $ref: "#/components/schemas/RiesgoControlDetalle"
            monitoreos:
              type: array
              items:
                $ref: "#/components/schemas/Monitoreo"

    # ── CONTROL ───────────────────────────────────────────────────────────────
    Control:
      type: object
      required: [id, codigo, descripcion, clase, tipo, frecuencia, formalidad, ponderacion, createdAt, updatedAt]
      properties:
        id: { type: integer }
        codigo: { type: string }
        descripcion: { type: string }
        clase: { type: string }
        pesoClase: { type: ["number", "null"] }
        tipo: { type: string }
        pesoTipo: { type: ["number", "null"] }
        frecuencia: { type: string }
        pesoFrecuencia: { type: ["number", "null"] }
        formalidad: { type: string }
        pesoFormalidad: { type: ["number", "null"] }
        soporte: { type: ["string", "null"] }
        responsable: { type: ["string", "null"] }
        ponderacion: { type: number }
        areas:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    RiesgoControlInput:
      type: object
      required: [controlId]
      properties:
        controlId: { type: integer }
        orden: { type: integer }
    RiesgoControlDetalle:
      type: object
      required: [controlId, riesgoId, orden]
      properties:
        controlId: { type: integer }
        riesgoId: { type: integer }
        orden: { type: integer }
        control:
          $ref: "#/components/schemas/Control"

    # ── MONITOREO ─────────────────────────────────────────────────────────────
    Monitoreo:
      type: object
      required: [id, codigo, aspectoMonitorear, riesgoId, createdAt, updatedAt]
      properties:
        id: { type: integer }
        codigo: { type: string }
        aspectoMonitorear: { type: string }
        indicador: { type: ["string", "null"] }
        periodicidad: { type: ["string", "null"] }
        responsable: { type: ["string", "null"] }
        riesgoId: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Parametro:
      type: object
      required: [id, categoria, nombre, valor, createdAt]
      properties:
        id: { type: integer }
        categoria: { type: string }
        nombre: { type: string }
        valor: { type: number }
        descripcion: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ParametroInput:
      type: object
      required: [categoria, nombre, valor]
      properties:
        categoria: { type: string }
        nombre: { type: string }
        valor: { type: number }
        descripcion: { type: string }
    ParametroUpdate:
      type: object
      properties:
        categoria: { type: string }
        nombre: { type: string }
        valor: { type: number }
        descripcion: { type: string }

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...