# workspace/lib/api-zod/src/generated/types/parametro.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API for Matriz de Riesgos LAFT
 * OpenAPI spec version: 0.1.0
 */

export interface Parametro {
  id: number;
  categoria: string;
  nombre: string;
  valor: number;
  /** @nullable */
  descripcion?: string | null;
  createdAt: Date;
}