# workspace/lib/api-zod/src/generated/types/parametroInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API for Matriz de Riesgos LAFT
 * OpenAPI spec version: 0.1.0
 */

export interface ParametroInput {
  categoria: string;
  nombre: string;
  valor: number;
  descripcion?: string;
}