# workspace/lib/api-zod/src/generated/types/riesgoControlDetalle.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API for Matriz de Riesgos LAFT
 * OpenAPI spec version: 0.1.0
 */
import type { Control } from './control';

export interface RiesgoControlDetalle {
  controlId: number;
  riesgoId: number;
  orden: number;
  control?: Control;
}