# workspace/lib/api-zod/src/generated/types/riesgoUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API for Matriz de Riesgos LAFT
 * OpenAPI spec version: 0.1.0
 */

export interface RiesgoUpdate {
  codigo?: string;
  proceso?: string;
  subproceso?: string;
  factorRiesgo?: string;
  descripcion?: string;
  riesgoLaft?: boolean;
  riesgoOperativo?: boolean;
  riesgoLegal?: boolean;
  riesgoReputacional?: boolean;
  riesgoContagio?: boolean;
  quePuedeSuceder?: string;
  tipologia?: string;
  porQuePuedeSuceder?: string;
  consecuencias?: string;
  probabilidadInherente?: number;
  impactoInherente?: number;
  probabilidadResidual?: number;
  impactoResidual?: number;
  tipoMonitoreo?: string;
  responsableMonitoreo?: string;
  periodicidadMonitoreo?: string;
  prioridad?: string;
  sugerencias?: string;
}